+++ doc/source/changelog.rst
@@ -5,6 +5,7 @@ Change Log
 v1.1.0
 ====================
 
+* Added `Support batch rendering & glob patterns in template selectors <https://github.com/bckohan/django-render-static/issues/15>`_
 * Fixed `Rename render_static -> renderstatic <https://github.com/bckohan/django-render-static/issues/11>`_
 * Added `Allow 'lazy' contexts built after Django bootstrapping <https://github.com/bckohan/django-render-static/issues/6>`_
 * Added `Flexible context specifiers <https://github.com/bckohan/django-render-static/issues/17>`_